sample_project/requirements.txt
scrapy>=2.5.0
scrapy-playwright>=0.0.16
pymongo>=4.0.0
python-dotenv>=0.19.0
itemadapter>=2.0.0
